+++ OpenGL framework/src/misc/util.cpp
@@ -5,7 +5,7 @@ std::vector<std::string> split_str(const std::string &s, char delim)
 	std::vector<std::string> elems;
 
 	const char* cstr = s.c_str();
-	const unsigned int strLength = s.length();
+	const unsigned int strLength = (unsigned int)s.length();
 	unsigned int start = 0;
 	unsigned int end = 0;
